backend/requirements.txt
# Flask & Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1


# Firebase
firebase-admin==6.2.0
pyjwt==2.8.0

# Vector Database
qdrant-client==1.16.2

# Graph Database
neo4j==5.15.0

# Search
elasticsearch==8.11.0

# AI/ML
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.2
spacy==3.7.2
numpy==1.26.2
scikit-learn==1.3.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Background Tasks
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
email-validator==2.1.0
phonenumbers==8.13.26
requests==2.31.0
httpx==0.25.2

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
